perf: optimize tag filtering and bitmap operations across multiple modules
- Remove unnecessary references in server.rs and shell/mod.rs for cleaner API usage
- Implement bulk bitmap operations in database.rs for improved tag counting performance
- Add fast bitmap count method in frame_ops.rs using range counting
- Pre-compute lowercase tags in lazy_tag_filter.rs to eliminate repeated allocations
- Optimize shard filtering in streamer.rs with improved state remapping
- Simplify tag index format v2 with direct UTF-8 strings and add bulk operations

+++ src/multi_list/frame_ops.rs
@@ -192,7 +192,7 @@ where
             if ub >= lb {
                 let count = if use_bitmaps {
                     if let Some(ref bm) = shard_infos[shard_idx].bitmap {
-                        bm.range(lb as u32..=ub as u32).count() as usize
+                        bm.range(lb as u32..=ub as u32).count()
                     } else {
                         0
                     }
